Began splitting functions into their own files (draw and label done)
